+++ waves/wcore/api/v2/views/services.py
@@ -51,7 +51,7 @@ class ServiceViewSet(viewsets.ReadOnlyModelViewSet):
 
     def get_queryset(self):
         """ retrieve available services for current request user """
-        return Service.objects.get_api_services(user=self.request.user)
+        return Service.objects.get_services(user=self.request.user)
 
     def list(self, request, **kwargs):
         """ List all available services """